requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
